+++ src/mafs/hitable.rs
@@ -3,14 +3,14 @@
 
 #[derive(Copy, Clone, Debug)]
  pub struct HitRecord{
-    pub t: f32,
+    pub t: f64,
     pub p: Vec3,
     pub normal: Vec3,
    //  pub front_face: bool
  }
 
  impl HitRecord{
-    pub fn new(t:f32 ,p: Vec3, normal: Vec3) -> Self{
+    pub fn new(t:f64 ,p: Vec3, normal: Vec3) -> Self{
         return HitRecord { t: 0.0, 
                            p: Vec3::new(0.0,0.0,0.0),
                            normal: Vec3::new(0.0,0.0,0.0)
@@ -19,5 +19,5 @@
 
  }
  pub trait Hitable{
-    fn hit(&self, t: Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
+    fn hit(&self, t: Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
  }
